requirements.txt
flask
faiss-cpu
pymupdf
openai
tqdm
numpy
python-dotenv
requests
pydantic
gunicorn